// VIS_Application/app/Model/Notification/TicketListOpen.ts
export interface ITicketListOpen {
    CompanyId: number;
    Mode: string;
    Id: number;
    Subject: string;
    CreatedOn: Date;
    AddressToGroup: string;
    AssignTo: string;
    CreatedByName: string;
    UpdatedByName: string;
    UpdatedOn: Date;
    Priority: string;
    Status: string;
    CreatedBy: number;
    UpdatedBy: number
    IsActive: boolean;
    UserId: number;
    FileName: string;

}